Move functions.
